analysis/requirements.txt
# Analysis dependencies for NeuroNav engagement analysis
pymongo>=4.0.0
pandas>=1.5.0
numpy>=1.20.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.9.0